+++ Example/Sources/Models/MockMessage.swift
@@ -25,6 +25,7 @@
 import Foundation
 import CoreLocation
 import MessageKit
+import AVFoundation
 
 private struct CoordinateItem: LocationItem {
 
@@ -49,6 +50,22 @@ private struct ImageMediaItem: MediaItem {
         self.image = image
         self.size = CGSize(width: 240, height: 240)
         self.placeholderImage = UIImage()
+    }
+
+}
+
+private struct MockAudiotem: AudioItem {
+
+    var url: URL
+    var size: CGSize
+    var duration: Float
+
+    init(url: URL) {
+        self.url = url
+        self.size = CGSize(width: 160, height: 35)
+        // compute duration
+        let audioAsset = AVURLAsset(url: url)
+        self.duration = Float(CMTimeGetSeconds(audioAsset.duration))
     }
 
 }
@@ -96,6 +113,11 @@ internal struct MockMessage: MessageType {
 
     init(emoji: String, sender: Sender, messageId: String, date: Date) {
         self.init(kind: .emoji(emoji), sender: sender, messageId: messageId, date: date)
+    }
+
+    init(audioURL: URL, sender: Sender, messageId: String, date: Date) {
+        let audioItem = MockAudiotem(url: audioURL)
+        self.init(kind: .audio(audioItem), sender: sender, messageId: messageId, date: date)
     }
 
 }